[feat]: 마인드맵 관련 상태 Zustand로 분리 및 적용
- 기존 mindmapStore 이름 변경 -> mindmapStore_deprecated
- 마인드맵 레이아웃 계산 관련 로직도 Store 코드에 분리
