+++ frontend/src/components/mindmap/MindmapWrapper.tsx
@@ -1,177 +1,28 @@
-import { useCallback, useState } from 'react';
 import {
   ReactFlow,
   ConnectionLineType,
   Panel,
-  useNodesState,
-  useEdgesState,
   ReactFlowProvider,
-  Node,
-  Edge,
-  NodeMouseHandler,
-  Position,
+  Controls,
+  NodeTypes,
 } from '@xyflow/react';
-import dagre from '@dagrejs/dagre';
-
 import '@xyflow/react/dist/style.css';
-import {
-  initialEdges,
-  initialNodes,
-} from '@/components/mindmap/initailElements';
+import CustomNode from '@/components/mindmap/CustomNode';
+import { Button } from '@/components/ui/button';
+import { DialogClose } from '@radix-ui/react-dialog';
+import { Modal } from '@/components/common/Modal';
+import { useMindmapStore } from '@/store/mindMapStore';
+
+const nodeTypes: NodeTypes = {
+  custom: CustomNode,
+};
 
 const flowStyles = {
   background: '#E8EFFF',
 };
 
-const nodeStyles = {
-  transition: 'all 0.5s ease',
-};
-
-const edgeStyles = {
-  transition: 'all 0.5s ease',
-  strokeWidth: 1.5,
-  stroke: '#0080ff',
-};
-
-const dagreGraph = new dagre.graphlib.Graph().setDefaultEdgeLabel(() => ({}));
-
-const nodeWidth = 172;
-const nodeHeight = 36;
-
-// 루트 노드 ID - 추후 삭제
-const ROOT_NODE_ID = '4';
-
-const getLayoutedElements = (
-  nodes: Node[],
-  edges: Edge[],
-  direction: 'TB' | 'LR' = 'TB',
-) => {
-  const isHorizontal = direction === 'LR';
-
-  // 루트 노드의 현재 위치 저장
-  const rootNode = nodes.find((node) => node.id === ROOT_NODE_ID);
-  const rootPosition = rootNode ? { ...rootNode.position } : { x: 0, y: 0 };
-
-  // Dagre를 사용하여 전체 그래프 레이아웃 계산
-  dagreGraph.setGraph({ rankdir: direction });
-
-  nodes.forEach((node) => {
-    dagreGraph.setNode(node.id, { width: nodeWidth, height: nodeHeight });
-  });
-
-  edges.forEach((edge) => {
-    dagreGraph.setEdge(edge.source, edge.target);
-  });
-
-  dagre.layout(dagreGraph);
-
-  // 루트 노드의 Dagre 계산 위치 가져오기
-  const dagreeRootPos = dagreGraph.node(ROOT_NODE_ID);
-  if (!dagreeRootPos) {
-    return { nodes, edges };
-  }
-
-  // Dagre가 계산한 루트 위치와 실제 루트 위치 간의 차이 계산
-  const diffX = rootPosition.x - (dagreeRootPos.x - nodeWidth / 2);
-  const diffY = rootPosition.y - (dagreeRootPos.y - nodeHeight / 2);
-
-  // 모든 노드에 대해 위치 조정 (트리 구조는 유지하면서 전체를 이동)
-  const newNodes = nodes.map((node) => {
-    const nodeWithPosition = dagreGraph.node(node.id);
-    if (!nodeWithPosition) {
-      return node;
-    }
-
-    // Dagre 계산 위치에 차이값을 더해 조정
-    const position = {
-      x: nodeWithPosition.x - nodeWidth / 2 + diffX,
-      y: nodeWithPosition.y - nodeHeight / 2 + diffY,
-    };
-
-    return {
-      ...node,
-      targetPosition: isHorizontal ? Position.Left : Position.Top,
-      sourcePosition: isHorizontal ? Position.Right : Position.Bottom,
-      position,
-    };
-  });
-
-  return { nodes: newNodes, edges };
-};
-
-// 초기 레이아웃 계산
-const { nodes: layoutedNodes, edges: layoutedEdges } = getLayoutedElements(
-  initialNodes,
-  initialEdges.map((edge) => ({
-    ...edge,
-    animated: true,
-    style: edgeStyles,
-  })),
-);
-
 function FlowContent() {
-  const [nodes, setNodes, onNodesChange] = useNodesState(layoutedNodes);
-  const [edges, setEdges, onEdgesChange] = useEdgesState(layoutedEdges);
-  const [currentDirection, setCurrentDirection] = useState<'TB' | 'LR'>('TB');
-
-  // 레이아웃 변경 시 부드러운 애니메이션 적용
-  const onLayout = useCallback(
-    (direction: 'TB' | 'LR') => {
-      setCurrentDirection(direction);
-      const { nodes: layoutedNodes, edges: layoutedEdges } =
-        getLayoutedElements(nodes, edges, direction);
-
-      setNodes(layoutedNodes);
-      setEdges(layoutedEdges);
-    },
-    [nodes, edges, setNodes, setEdges],
-  );
-
-  // 노드 클릭 핸들러 추가
-  const onNodeClick: NodeMouseHandler = useCallback(
-    (event, clickedNode) => {
-      // 랜덤 ID 생성
-      const randomId = Math.random().toString(36).substring(2, 10);
-      const newNodeId = `node-${randomId}`;
-
-      // 새 하위 노드 생성
-      const newChildNode: Node = {
-        id: newNodeId,
-        data: { label: `Child of ${clickedNode.id}` },
-        // 초기 위치는 클릭된 노드 아래에 배치 (정확한 위치는 어차피 재배치됨!)
-        position: {
-          x: clickedNode.position.x + 100,
-          y: clickedNode.position.y + 100,
-        },
-        targetPosition:
-          currentDirection === 'LR' ? Position.Left : Position.Top,
-        sourcePosition:
-          currentDirection === 'LR' ? Position.Right : Position.Bottom,
-        style: nodeStyles,
-      };
-
-      // 상위 노드와 하위 노드를 연결하는 새 엣지 생성
-      const newEdge: Edge = {
-        id: `edge-${clickedNode.id}-${newNodeId}`,
-        source: clickedNode.id,
-        target: newNodeId,
-        type: ConnectionLineType.SmoothStep,
-        animated: true,
-        style: edgeStyles,
-      };
-
-      // 새 노드와 엣지 추가 후 레이아웃 재계산
-      const updatedNodes = [...nodes, newChildNode];
-      const updatedEdges = [...edges, newEdge];
-
-      const { nodes: layoutedNodes, edges: layoutedEdges } =
-        getLayoutedElements(updatedNodes, updatedEdges, currentDirection);
-
-      setNodes(layoutedNodes);
-      setEdges(layoutedEdges);
-    },
-    [nodes, edges, currentDirection, setNodes, setEdges],
-  );
+  const { nodes, edges, onNodesChange, onEdgesChange } = useMindmapStore();
 
   return (
     <ReactFlow
@@ -179,27 +30,64 @@ function FlowContent() {
       edges={edges}
       onNodesChange={onNodesChange}
       onEdgesChange={onEdgesChange}
-      onNodeClick={onNodeClick}
-      connectionLineType={ConnectionLineType.SmoothStep}
+      nodeTypes={nodeTypes}
+      connectionLineType={ConnectionLineType.Bezier}
       fitView
       style={flowStyles}
       defaultViewport={{ x: 0, y: 0, zoom: 1 }}
       nodesDraggable={false}
     >
-      <Panel position="top-right">
-        <div className="flex gap-2">
-          <button
-            onClick={() => onLayout('TB')}
-            className="px-3 py-1 bg-blue-500 text-white rounded"
-          >
-            vertical layout
-          </button>
-          <button
-            onClick={() => onLayout('LR')}
-            className="px-3 py-1 bg-blue-500 text-white rounded"
-          >
-            horizontal layout
-          </button>
+      <Controls />
+
+      <Panel position="bottom-center">
+        <div className="mb-12 z-10 bg-[rgba(255,255,255,0.6)] rounded-4xl px-6 py-4 w-auto">
+          <div className="flex items-center gap-3">
+            <Modal
+              trigger={
+                <Button className="w-[139px] h-[48px] text-center rounded-4xl bg-blue-2 text-blue font-semibold">
+                  취소
+                </Button>
+              }
+              title="취소"
+              description="마인드맵 작성을 취소하시겠습니까?"
+              footer={
+                <div className="w-full flex items-center justify-end">
+                  <DialogClose asChild>
+                    <Button size="sm" className="bg-blue text-white">
+                      적용하기
+                    </Button>
+                  </DialogClose>
+                </div>
+              }
+            >
+              <div className="rounded-[7px] px-6 py-[20px] text-[20px] font-semibold bg-blue-2">
+                마인드맵 내용은 저장되지 않으며, 다시 확인할 수 없습니다.
+              </div>
+            </Modal>
+
+            <Modal
+              trigger={
+                <Button className="w-[139px] h-[48px] text-center rounded-4xl bg-blue text-white font-semibold">
+                  완료
+                </Button>
+              }
+              title="한 줄 요약"
+              description="마인드맵으로 정리된 한 줄 요약입니다."
+              footer={
+                <div className="w-full flex items-center justify-end">
+                  <DialogClose asChild>
+                    <Button size="sm" className="bg-blue text-white">
+                      적용하기
+                    </Button>
+                  </DialogClose>
+                </div>
+              }
+            >
+              <div className="rounded-[7px] px-6 py-[20px] text-[20px] font-semibold bg-blue-2">
+                마인드맵 개발을 위해 React Flow 라이브러리 공부하기
+              </div>
+            </Modal>
+          </div>
         </div>
       </Panel>
     </ReactFlow>
